+++ handlerbars/loader.js
@@ -1,0 +1,9 @@
+// Notice: load partials
+require('./partials/render-nested');
+
+// Notice: load helpers
+require('./helpers/equal');
+require('./helpers/indent');
+require('./helpers/is-array');
+require('./helpers/is-object');
+require('./helpers/sum');