+++ test-fixtures/mongo/nested-array-of-objects-model.js
@@ -1,0 +1,22 @@
+const { ObjectID } = require('mongodb');
+
+const persons = [
+  {
+    _id: ObjectID(),
+    name: 'James Cameron',
+    propArrayOfObjects: [
+      {
+        _id: ObjectID(),
+        one: 'one',
+        two: 'two',
+      },
+      {
+        _id: ObjectID(),
+        one: '1',
+        two: '2',
+      },
+    ],
+  },
+];
+
+module.exports = { persons };