+++ test-fixtures/mongo/sub-documents-not-using-ids-model.js
@@ -1,0 +1,18 @@
+const { ObjectID } = require('mongodb');
+
+const persons = [
+  {
+    _id: ObjectID(),
+    name: 'James Cameron',
+    propArrayOfObjects: [
+      {
+        sampleValue: 'sample',
+      },
+      {
+        sampleValue: 'sample',
+      },
+    ],
+  },
+];
+
+module.exports = { persons };